added response object

+++ src/main/java/com/jpop/bookservice/entity/Book.java
@@ -67,9 +67,9 @@ public class Book {
     private Date updatedAt;
 
     @Column(name = "created_by")
-    private Long createdBy;
+    private Integer createdBy;
 
     @Column(name = "updated_by")
-    private Long updatedBy;
+    private Integer updatedBy;
 
 }
